update harvard pilot code to not use a harvard id; substitute smail address for employeeid instead refactor integration tests to make them easier to maintain and understand
