+++ pass-grant-data/src/test/java/org/dataconservancy/pass/grant/data/HarvardPilotConnectorTest.java
@@ -101,7 +101,7 @@ public class HarvardPilotConnectorTest {
                 "C",
                 "P"
         };
-        String[] harvardIds = {"86753091",
+ /*       String[] harvardIds = {"86753091",
                 "86753092",
                 "86753093",
                 "86753094",
@@ -110,6 +110,16 @@ public class HarvardPilotConnectorTest {
                 "86753097",
                 "86753098",
                 "86753099"
+        };  commented out because we aren't using these (yet) - emails instead*/
+        String[] harvardIds = {"wdrumstick@harvard.edu",
+                "gflanksteak@harvard.edu",
+                "ddbrisket@harvard.edu",
+                "abacon@harvard.edu",
+                "efarmer@harvard.edu",
+                "",
+                "csteer@harvard.edu",
+                "dbovine@harvard.edu",
+                "bcow@harvard.edu"
         };
         String[] emails = {"wdrumstick@harvard.edu",
                 "gflanksteak@harvard.edu",
